server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-agents-server"
version = "0.1.0"
description = "Pixel Agents webapp backend"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "takopi",
    "anyio>=4.0",
    "watchfiles>=1.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.setuptools.packages.find]
where = [".."]
include = ["server*"]
